Include only the lib directory in the gem
All the files required for users to use this gem are contained in the
lib directory. The lib directory also only contains files that are
potentially necessary for users; that is, it has no unnecessary files.
Including only this lib directory keeps the gem size minimal.

See 'What Files to Include?' on https://piotrmurach.com/articles/writing-a-ruby-gem-specification/